move dashboard files
